🔊 Meaningful error message for sd version conflict (#2268)
* :truck: Move enums

* :loud_sound: Meaningful error message for sd version conflict
